# openapi.yaml
openapi: "3.0.3"
info:
  title: lab-res API
  version: "1"
  contact:
    name: Healthmetrix GmbH
    url: http://www.healthmetrix.com
    email: admin@healthmetrix.com
servers:
  - url: "https://api.labres.dev.healthmetrix.com/{version}"
    description: development environment for the LabRes apiKey
    variables:
      version: 
        default: v1
tags:
  - name: lab_api
    description: endpoints to be invoked from the lab to upload results
  - name: orders_api
    description: endpoints to be invoked by a backend to empower patients to query their lab results
components:
  securitySchemes:
    LabApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: API key authorizing access for labs to upload results for an external order number
    OrdersApiKey:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GeneralError:
      type: object
      properties:
        message:
          type: string
  responses:
    UnauthorizedApiKeyError:
      description: API key invalid or missing
      headers:
        WWW-Authenticate:
          schema:
            type: string
    GeneralError:
      description: General error response body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
paths:
  /orders:
    post:
      summary: issues a new order and returns the created external order number
      description: Should only be invoked for verified users (logged into account or verified email address)
      tags:
        - orders_api
      responses:
        '201':
          description: a new external order number to be returned
          content: 
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: A unique internal identifier for the order
                    format: UUID
                  externalOrderNumber:
                    type: string
                    format: externalOrderNumber
                    description: numeric 8-digit-long identifier
                  token:
                    type: string
                    description: Authentication Token
                example:
                  id: 65e524cb-7494-4073-ad16-495fed0d79e4
                  externalOrderNumber: X72UN3K8
                  token: "abc123"
        '401':
          $ref: "#/components/responses/UnauthorizedApiKeyError"
        '500':
          description: Internal server error
          $ref: "#/components/responses/GeneralError"
      security:
        - OrdersApiKey: []
  /orders/{orderId}:
    get:
      summary: returns the current status of a lab order
      description: Should only be invoked for verified users (logged into account or verified email address)
      tags:
        - orders_api
      parameters:
        - name: orderId
          in: path
          description: UUID of an order that has been sent to a lab
          required: true
          schema:
            type: string
            format: uuid
            description: unique identifier of a lab order (NOT the external order number)
      responses:
        '200':
          description: returns current status of the lab order
          content: 
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [POSITIVE, WEAK_POSITIVE, NEGATIVE, INVALID, IN_PROGRESS]
                    description: Status enum to indicate if the lab result is in progress, sample material has been found invalid,  or if it has a positive or negative result
                example:
                  status: POSITIVE
        '401':
          $ref: "#/components/responses/UnauthorizedApiKeyError"
        '404':
          description: no order for the given id found
        '500':
          description: Internal server error
          $ref: "#/components/responses/GeneralError"
      security:
        - OrdersApiKey: []
  /order/{externalOrderNumber}/result:
    put:
      summary: allows for uploading a result of a lab order
      tags:
        - lab_api
      parameters:
        - name: externalOrderNumber
          in: path
          description: external ID of an order that has been sent to a lab
          required: true
          schema:
            type: string
            format: externalOrderNumber
            description: numeric 8-digit-long identifier
      requestBody:
        description: Lab result as OBX message or result in JSON body
        content:
          'text/plain':
            schema:
              type: string
              format: OBX
              description: An OBX Segment of an HL7 ORU R1 message. See https://wiki.hl7.de/index.php?title=Segment_OBX for details.
          'application/json':
            schema:
              type: object
              properties:
                result:
                  type: string
                  description: The lab result as one of the following strings "Positiv", "Schwach positiv", "Nicht nachweisbar", "Prozessfehler".
              required:
                - result
      responses:
        '200':
          description: Result uploaded successfully
        '401':
          $ref: "#/components/responses/UnauthorizedApiKeyError"
        '404':
          description: no order for external order number found
        '500':
          description: Internal server error
          $ref: "#/components/responses/GeneralError"
      security:
        - LabApiKey: []
  /lab-orders/{labOrderNumber}:
    put:
      summary: Creates an Internal Order Number (ION)
      description: Should only be invoked for verified users (logged into account or verified email address)
      tags:
        - lab_api
      parameters:
        - name: labOrderNumber
          in: path
          description: Laboratory Order Number
          required: true
          schema:
            type: string
            format: alphanumeric identifier
      responses:
        '201':
          description: Lab Order Number created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: uuid identifier of the lab order number
                  labOrderNumber:
                    type: string
                    description: the lab order number just created
                example:
                  id: f8aeb720-23a8-42f5-9be2-176d672c13a7
                  labOrderNumber: 1234abcd
        '401':
          $ref: "#/components/responses/UnauthorizedApiKeyError"
        '500':
          description: Internal server error
          $ref: "#/components/responses/GeneralError"
      security:
        - LabApiKey: []
  /lab-orders/{orderId}:
    get:
        summary: Fetches a lab order's status
        description: Fetches an lab order's result status based on the order id (not the lab order number)
        tags:
          - lab_api
        parameters:
          - name: orderId
            in: path
            description: Lab Order ID
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: Lab Order Number found
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      enum: [POSITIVE, WEAK_POSITIVE, NEGATIVE, INVALID, IN_PROGRESS]
                      description: Status enum to indicate if the lab result is in progress, sample material has been found invalid,  or if it has a positive or negative result
                  example:
                    status: POSITIVE
          '401':
            $ref: "#/components/responses/UnauthorizedApiKeyError"
          '500':
            description: Internal server error
            $ref: "#/components/responses/GeneralError"
        security:
          - LabApiKey: []